rnrrn

+++ GreedyAnt.h
@@ -5,7 +5,7 @@
 class GreedyAnt : public Ant
 {
 public:
-    int nextLibrary(std::vector<Library> &libraries, int iter, int p) override;
+    int nextLibrary(std::vector<Library> &libraries, int iter, double p) override;
     GreedyAnt(int deadline, int numberOfLibraries);
 };
